% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.2_parametric.R
\name{BoxCox}
\alias{BoxCox}
\title{using a Box-Cox transformation to transform continuous data into binormal data}
\usage{
BoxCox(data0, data1)
}
\arguments{
\item{data0}{continuous test results of nondiseased individuals}

\item{data1}{continuous test results of diseased individuals}
}
\value{
the optimal parameter value for the Box-Cox transformation,the binormal parameters after transformation and corresponding AUC
#example
#data1<-c(140,1087,230,183,1256,700,16,800,253,740,126,153,283,90,303,193,76,1370,543,913,230,463,60,509,576,671,80,490,156,356,350,323,1560,120,216,443,523,76,303,353,206)
#data0<-c(136,286,281,23,200,146,220,96,100,60,17,27,126,100,253,70,40,6,46)
#BoxCox(data0,data1)
}
\description{
using a Box-Cox transformation to transform continuous data into binormal data
}
