% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.4.R
\name{CI_estimation_continuous}
\alias{CI_estimation_continuous}
\title{function to estimate confidence intervals for full area and partial area under binormal assumption}
\usage{
CI_estimation_continuous(data0, data1, e1, e2, alpha)
}
\arguments{
\item{data0}{continuous test results of nondiseased subjects}

\item{data1}{continuous test results of diseased subjects}

\item{e1}{lower bound of FPR for partial area}

\item{e2}{upper bound of FPR for partial area}

\item{alpha}{significance level}
}
\value{
CI_full,CI_partial:directly using Wald's interval
        CI_full_logit,CI_partial_logit:using logit transformation together with Wald's interval
        CI_partial_McClish:using McClish's transformation
#example
#data1<-c(140,1087,230,183,1256,700,16,800,253,740,126,153,283,90,303,193,76,1370,543,913,230,463,60,509,576,671,80,490,156,356,350,323,1560,120,216,443,523,76,303,353,206)
#data0<-c(136,286,281,23,200,146,220,96,100,60,17,27,126,100,253,70,40,6,46)
#CI_estimation_continuous(data0,data1,0,0.2,0.05)
}
\description{
function to estimate confidence intervals for full area and partial area under binormal assumption
}
