% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.5.R
\name{CI_estimation_ordinal}
\alias{CI_estimation_ordinal}
\title{function to estimate confidence intervals for full area and partial area}
\usage{
CI_estimation_ordinal(data0, data1, e1, e2, alpha)
}
\arguments{
\item{data0}{ordinal test results of nondiseased subjects}

\item{data1}{ordinal test results of diseased subjects}

\item{e1}{lower bound of FPR for partial area}

\item{e2}{upper bound of FPR for partial area}

\item{alpha}{significance level}
}
\value{
CI_full,CI_partial:directly using Wald's interval
        CI_full_logit,CI_partial_logit:using logit transformation together with Wald's interval
        CI_partial_McClish:using McClish's transformation
#example
#data0<-c(38,25,15,19,4)
#data1<-c(1,2,3,14,42)
#CI_estimation_ordinal(data0,data1,0,0.2,0.05)
}
\description{
function to estimate confidence intervals for full area and partial area
}
