% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.2_nonparametric.R
\name{Epanechnikov_kernel}
\alias{Epanechnikov_kernel}
\title{kernel function proposed by Zhou and Herezlar}
\usage{
Epanechnikov_kernel(t)
}
\arguments{
\item{t}{calculate the kernel}
}
\description{
kernel function proposed by Zhou and Herezlar
}
