% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.2_semiparametric.R
\name{getconstant}
\alias{getconstant}
\title{Function used to fitting a smooth ROC curve.It is not important.}
\usage{
getconstant(data0, data1)
}
\arguments{
\item{data0}{continuous test result of undiseased subjects}

\item{data1}{continuous test result of diseased subjects}
}
\description{
Function used to fitting a smooth ROC curve.It is not important.
}
