% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.4.R
\name{LRratio}
\alias{LRratio}
\title{Compute positive and negative LR as well as their confidence limits for binary-scale tests}
\usage{
LRratio(data, alpha)
}
\arguments{
\item{data}{2*2 table of test results of diseased and undiseased subjects}

\item{alpha}{significance level}
}
\value{
Estimation of positive and negative LR,Estimation of the variance of logarithms of positive and negative LR,
Estimation of confidence limits for positive and negative LR
#example
#data<-matrix(c(22,3,2,3),nrow=2,byrow=TRUE)
#LRratio(data,0.05)
}
\description{
Compute positive and negative LR as well as their confidence limits for binary-scale tests
}
