% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.5.R
\name{Oddsratio}
\alias{Oddsratio}
\title{Compute Odds ratio and its confidence limits for binary-scale tests}
\usage{
Oddsratio(data, alpha)
}
\arguments{
\item{data}{2*2 table of test results of diseased and nondiseased subjects}

\item{alpha}{significance level}
}
\value{
estimation of odds ratio and its confidence limits
#example
#data<-matrix(c(22,3,2,3),nrow=2,byrow=TRUE)
#Oddsratio(data,0.05)
}
\description{
Compute Odds ratio and its confidence limits for binary-scale tests
}
