% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.2-4.2.4.R
\name{params_estimate}
\alias{params_estimate}
\title{Function to estimate the binormal parameters a and b for ordinal data}
\usage{
params_estimate(data0, data1)
}
\arguments{
\item{data0}{test results of nondiseased subjects}

\item{data1}{test results of diseased subjects}
}
\value{
estimation of all thresholds,the mean and sd for undiseased subjects,the mean and sd for diseased subjects,and the covariance matrix for binormal parameters a and b
#example
#data0<-c(38,25,15,19,4)
#data1<-c(1,2,3,14,42)
#params_estimate(data0,data1)
}
\description{
Function to estimate the binormal parameters a and b for ordinal data
}
