% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.8.R
\name{params_estimate_bigamma}
\alias{params_estimate_bigamma}
\title{function to estimate the parameters correlated to bigamma model}
\usage{
params_estimate_bigamma(data0, data1)
}
\arguments{
\item{data0}{ordinal test results of nondiseased subjects}

\item{data1}{ordinal test results of diseased subjects}
}
\value{
estimation of all thresholds,the bigamma parameters alpha and sigma for nondiseased subjects,the bigamma parameters alpha and sigma for diseased subjects
#example
#data0<-c(38,25,15,19,4)
#data1<-c(1,2,3,14,42)
#params_estimate_bigamma(data0,data1)
}
\description{
function to estimate the parameters correlated to bigamma model
}
