% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.8.R
\name{params_estimate_logistic}
\alias{params_estimate_logistic}
\title{function to estimate the parameters correlated to logistic model}
\usage{
params_estimate_logistic(data0, data1)
}
\arguments{
\item{data0}{ordinal test results of nondiseased subjects}

\item{data1}{ordinal test results of diseased subjects}
}
\value{
estimation of all thresholds,the logistic parameters mu and gamma for nondiseased subjects,the logistic parameters mu and gamma for diseased subjects
#example
#data0<-c(38,25,15,19,4)
#data1<-c(1,2,3,14,42)
#params_estimate_logistic(data0,data1)
}
\description{
function to estimate the parameters correlated to logistic model
}
