% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.4.R
\name{Psi}
\alias{Psi}
\title{this function is used to calculate AUC for continuous data.It is not important.}
\usage{
Psi(x, y)
}
\arguments{
\item{x}{not important}

\item{y}{not important}
}
\description{
this function is used to calculate AUC for continuous data.It is not important.
}
