% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.4.R
\name{roc.area.nonparametric}
\alias{roc.area.nonparametric}
\title{function used to calculate AUC and partial AUC for continuous data using nonparametric methods}
\usage{
roc.area.nonparametric(data0, data1, e1, e2)
}
\arguments{
\item{data0}{continuous test result for undiseased subjects}

\item{data1}{continuous test result for diseased subjects}

\item{e1}{lower limit of FPR for partial AUC}

\item{e2}{upper limit of FPR for partial AUC}
}
\value{
ustimation of AUC,partial AUC and its variance using Dodd's and He's methods
#example
#data1<-c(140,1087,230,183,1256,700,16,800,253,740,126,153,283,90,303,193,76,1370,543,913,230,463,60,509,576,671,80,490,156,356,350,323,1560,120,216,443,523,76,303,353,206)
#data0<-c(136,286,281,23,200,146,220,96,100,60,17,27,126,100,253,70,40,6,46)
#roc.area.nonparametric(data0,data1,0,0.2)
}
\description{
function used to calculate AUC and partial AUC for continuous data using nonparametric methods
}
