% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.1.R
\name{roc.binary}
\alias{roc.binary}
\title{Compute Sensitivity and Specificity for binary scale tests}
\usage{
roc.binary(data, alpha)
}
\arguments{
\item{data}{2*2 table of test results of diseased and nondiseased subjects}

\item{alpha}{significance level}
}
\value{
estimation of sensitivity and specificity,their variances,and the estimated
confidence intervals,including wald's,AC and ZL intervals.
#example
#data<-matrix(c(22,3,2,3),nrow=2,byrow=TRUE)
#roc.binary(data,0.05)
}
\description{
Compute Sensitivity and Specificity for binary scale tests
}
