% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.7.R
\name{roc.binormality_optimal}
\alias{roc.binormality_optimal}
\title{FPR of the optimal decision point when assuming binormality}
\usage{
roc.binormality_optimal(data0, data1, m)
}
\arguments{
\item{data0}{continuous test results of nondiseased subjects}

\item{data1}{continuous test results of diseased subjects}

\item{m}{it equals R*(1-p)/p,where p is prevalence rate and R is ratio of the cost}
}
\value{
the optimal FPR and the corresponding TPR and decision threshold
#example
#data1<-c(140,1087,230,183,1256,700,16,800,253,740,126,153,283,90,303,193,76,1370,543,913,230,463,60,509,576,671,80,490,156,356,350,323,1560,120,216,443,523,76,303,353,206)
#data0<-c(136,286,281,23,200,146,220,96,100,60,17,27,126,100,253,70,40,6,46)
#roc.binormality_optimal(data0,data1,1)
}
\description{
FPR of the optimal decision point when assuming binormality
}
