% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.3.R
\name{roc.clusteredbinary}
\alias{roc.clusteredbinary}
\title{Compute Sensitivity, Specificity and Predictive Values with Clustered Binary-Scale Data}
\usage{
roc.clusteredbinary(data, alpha)
}
\arguments{
\item{data}{a 3-dimensional array
each cluster is represented by a 2*2 table(a matrix)
the third dimension is the number of cluster(the number of matrices)}

\item{alpha}{significance level}
}
\value{
estimation of Se,Sp,PPV,NPV,their variances and confidence intervals(wald's interval)
#example
#matrix1<-matrix(c(2,0,0,1),nrow=2,byrow=TRUE)
#matrix2<-matrix(c(1,1,1,0),nrow=2,byrow=TRUE)
#matrix3<-matrix(c(1,0,0,1),nrow=2,byrow=TRUE)
#matrix4<-matrix(c(1,0,0,2),nrow=2,byrow=TRUE)
#matrix5<-matrix(c(0,1,1,0),nrow=2,byrow=TRUE)
#matrix6<-matrix(c(2,0,1,1),nrow=2,byrow=TRUE)
#matrix7<-matrix(c(1,0,0,1),nrow=2,byrow=TRUE)
#matrix8<-matrix(c(2,1,0,1),nrow=2,byrow=TRUE)
#matrix9<-matrix(c(0,1,1,1),nrow=2,byrow=TRUE)
#matrix10<-matrix(c(1,1,0,1),nrow=2,byrow=TRUE)
#data<-array(c(matrix1,matrix2,matrix3,matrix4,matrix5,matrix6,matrix7,matrix8,matrix9,matrix10),dim=c(2,2,10))
#roc.clusteredbinary(data,0.05)
}
\description{
Compute Sensitivity, Specificity and Predictive Values with Clustered Binary-Scale Data
}
