% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.1.R
\name{roc.continuous.emp}
\alias{roc.continuous.emp}
\title{draw an empirical ROC curve for continuous data}
\usage{
roc.continuous.emp(data0, data1)
}
\arguments{
\item{data0}{test results of undiseased subjects}

\item{data1}{test results of diseased subjects}
}
\value{
an empirical ROC curve
#example
#data1<-c(140,1087,230,183,1256,700,16,800,253,740,126,153,283,90,303,193,76,1370,543,913,230,463,60,509,576,671,80,490,156,356,350,323,1560,120,216,443,523,76,303,353,206)
#data0<-c(136,286,281,23,200,146,220,96,100,60,17,27,126,100,253,70,40,6,46)
#roc.continuous.emp(data0,data1)
}
\description{
draw an empirical ROC curve for continuous data
}
