% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.8.R
\name{roc.logistic}
\alias{roc.logistic}
\title{function to plot the ROC curve under logistic assumption}
\usage{
roc.logistic(data0, data1)
}
\arguments{
\item{data0}{test results of nondiseased subjects}

\item{data1}{test results of diseased subjects}
}
\value{
a smooth ROC curve
#example
#data0<-c(38,25,15,19,4)
#data1<-c(1,2,3,14,42)
#roc.logistic(data0,data1)
}
\description{
function to plot the ROC curve under logistic assumption
}
