% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.2-4.2.4.R
\name{roc.ordinal.smooth}
\alias{roc.ordinal.smooth}
\title{Function to fully analyze binormal data}
\usage{
roc.ordinal.smooth(data0, data1, alpha, FPR, e1, e2)
}
\arguments{
\item{data0}{test results of nondiseased subjects}

\item{data1}{test results of diseased subjects}

\item{alpha}{significance level}

\item{FPR}{estimate TPR at a particular FPR rate}

\item{e1}{lower bound of FPR rate for partial area}

\item{e2}{upper bound of FPR rate for partial area}
}
\value{
results of optimization algorithm,including decision thresholds and binormal parameters.
        binormal_indicator:it equals one if the goodness-of-fit test is passed.
        threshold_indicator:it equals one if Zhou's test is passed.
        nondiseased,diseased:mean and standard deviation of latent normal variable for nondiseased and diseased subgroup
        a,b,var_a,var_b,covar_ab:estimated binormal paramaters,their variances and covariances
        TPR,TPR_CI:estimated sensitivity at the fixed FPR and its confidence interval
        area.full,area.partial:estimated full area and partial area
        var_full,var_partial:direct estimation of the variances of full area and partial area
        partial_transformed,var_transformed:estimation of partial area and its variance using McClish transformation
#example
#data0<-c(38,25,15,19,4)
#data1<-c(1,2,3,14,42)
#roc.ordinal.smooth(data0,data1,0.05,0.1,0,0.2)
}
\description{
Function to fully analyze binormal data
}
