% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.4.R
\name{roc.testing}
\alias{roc.testing}
\title{testing the hypothesis that the roc curve area or partial area is a specific value}
\usage{
roc.testing(A0, A, variance, alpha)
}
\arguments{
\item{A0}{area or partial area under the null hypothesis}

\item{A}{the estimated value of area or partial area}

\item{variance}{estimated variance of area or partial area}

\item{alpha}{significance level}
}
\value{
whether the null hypothesis is rejected
#example
#roc.testing(0.5,0.9297,0.00045369,0.05)
}
\description{
testing the hypothesis that the roc curve area or partial area is a specific value
}
