% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.3.2_semiparametric.R
\name{star}
\alias{star}
\title{Function used to fit a smooth ROC curve.It is not important.}
\usage{
star(Y_star, k_star, l_star)
}
\arguments{
\item{Y_star}{not important}

\item{k_star}{not important}

\item{l_star}{not important}
}
\description{
Function used to fit a smooth ROC curve.It is not important.
}
