% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.7.R
\name{roc.clustered}
\alias{roc.clustered}
\title{using nonparametric method to estimate AUC and its variance}
\usage{
roc.clustered(data)
}
\arguments{
\item{data}{the test result.It is a matrix consisting of even rows
each cluster corresponds to two rows
The first row is the ordinal test results for diseased units in that cluster
The second row is the ordinal test results for undiseased units in that cluster
The cols represent ordinal test results.}
}
\value{
estimation of AUC and its variance
#example
#data<-matrix(c(1,2,3,14,42,38,25,15,19,4,8,2,2,2,48,70,7,5,7,12),nrow=4,byrow=TRUE)
#roc.clustered(data)
}
\description{
using nonparametric method to estimate AUC and its variance
}
