% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.1.3.R
\name{roc.clusteredbinary}
\alias{roc.clusteredbinary}
\title{Compute Sensitivity, Specificity and Predictive Values with Clustered Binary-Scale Data}
\usage{
roc.clusteredbinary(data, alpha)
}
\arguments{
\item{data}{a 3-dimensional array
each cluster is represented by a 2*2 table(a matrix)
the third dimension is the number of cluster(the number of matrices)}

\item{alpha}{significance level}
}
\value{
estimation of Se,Sp,PPV,NPV,their variances and confidence intervals(wald's interval)
#example
#matrix1 <- matrix(c(22, 3, 2, 3), nrow = 2, byrow = TRUE)
#matrix2 <- matrix(c(56, 6, 23, 78), nrow = 2, byrow = TRUE)
#data <- array(c(matrix1, matrix2), dim = c(2, 2, 2))
#roc.clusteredbinary(data,0.05)
}
\description{
Compute Sensitivity, Specificity and Predictive Values with Clustered Binary-Scale Data
}
