% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.2.6.R
\name{roc.nonparametric}
\alias{roc.nonparametric}
\title{using nonparametric method to estimate full area and partial area under ROC curve
also estimate the variance of full area}
\usage{
roc.nonparametric(data0, data1, p, s)
}
\arguments{
\item{data0}{test results of nondiseased subjects}

\item{data1}{test results of diseased subjects}

\item{p}{the partial area between the FPRs e1 and e2, which correspond to the p-th and s-th ordinal test results}

\item{s}{the partial area between the FPRs e1 and e2, which correspond to the p-th and s-th ordinal test results}
}
\value{
estimation of full area using MW estimate and Delong's estimate and their variances,
estimation of partial area
#example
#data0<-c(38,25,15,19,4)
#data1<-c(1,2,3,14,42)
#roc.nonparametric(data0,data1,2,4)
}
\description{
using nonparametric method to estimate full area and partial area under ROC curve
also estimate the variance of full area
}
